Split out test_engine_download.py.
